feat(geo-index): Refactor library to use quadtree index for geographic lookup of tz data
This refactor aims to improve the speed and lower the memory usage of this library.  Before, all of

the world's tz data was loaded into memory and then every single tz was searched until there was a

match.  This refactor includes automatic generation of a quadtree index of the tz data although

still searches in much smaller files to determine the exact timezone if needed.