fix: remove default export in favor of named exports
Fixes #129

microbundle is not happy with the mixing of default and named exports. It seems the easiest thing to do here is make the default export a named export.

This is probably a breaking change from v7.0.0, but since it was released not long ago and not really working all that well anyways, I'm just calling this a fix and assuming this named export thing should've been a breaking change that landed in v7.0.0.

This is probably a violation of SEMVER in a strict sense, so I apologize to anyone who already made workarounds for what was in v7.0.0.